Improved globbing behavior.
Previously, we'd try to glob any filepath. This had the unintentional
consequence that missing files that should have been flagged as errors
would not be tested since the glob match didn't find them in the first
place. In the updated code we only glob if an "*" glob pattern is present
and even then we check that we at least get a single match (i.e. empty
glob patterns are not allowed).