Added some preliminary dynamic geometry tests.
Test if molecules move into the appropriate mesh after a dynamic geometry
event. Specifically, this makes sure volume molecules respect rules regarding
transparencies.

In two of the tests, there are three nested cubes and several species of volume
molecules.

In the first test case, the volume molecules (reperesented below with an
asterisk) start outside of the cubes and move such that the molecules end up
inside the cubes.

     --------------         --------------
     |   _______  |         |   _______  |
     |  |   _   | |         |  |   _   | |
  *  |  |  | |  | |  --->   |  |  |*|  | |
     |  |_______| |         |  |_______| |
     |            |         |            |
     --------------         --------------

In the other case, the volume molecules start inside the cubes and move such
that the molecules end up outside them.

     --------------         --------------
     |   _______  |         |   _______  |
     |  |   _   | |         |  |   _   | |
     |  |  |*|  | |  --->   |  |  | |  | |*
     |  |_______| |         |  |_______| |
     |            |         |            |
     --------------         --------------

The third and final case tests movement between two sets of "non-overlapping"
cubes.

   --------------   --------------       --------------   --------------
   |   _______  |   |   _______  |       |   _______  |   |   _______  |
   |  |   _   | |   |  |   _   | |       |  |   _   | |   |  |   _   | |
   |  |  |*|  | |   |  |  | |  | |  -->  |  |  | |  | |   |  |  |*|  | |
   |  |_______| |   |  |_______| |       |  |_______| |   |  |_______| |
   |            |   |            |       |            |   |            |
   --------------   --------------       --------------   --------------

Just because the meshes are moved (either around or away from molecules),
doesn't mean the molecules will stay in the new location. Only species that are
transparent to those surfaces should stay where they are. This means some
molecules will have to be moved in or out of various meshes.