Test to add/remove objects during dyngeom events.
This test makes sure that MCell behaves sensibly and doesn't crash when
we add or remove an object during a dynamic geometry event. The policy
decision for now is that MCell should ignore objects that were either
just added or removed for the current dynamic geometry event when
placing molecules.