Test to add/remove regions during dyngeom events.
This test makes sure that MCell behaves sensibly and doesn't crash when we add
or remove a region during a dynamic geometry event. In this particular test
case, we release molecules on sr1 (top of cube). Then region sr2 is removed and
sr3 is added during the dynamic geometry event. The policy decision for now is
that MCell should ignore objects that were either just added or removed for the
current dynamic geometry event when placing molecules.