Add new DG test for reactions and counts.
Two nearly identical models are included in this test. The only
difference is that the one uses dynamic geometries and the other
doesn't. To keep things as similar as possible, the same geometry is
used for every single DG event, which is also the same geometry used in
the non-DG case.

Three basic reactions are tested: two different bimolecular reactions
(one using surface molecules and the other volume molecules) and a
unimolecular reaction.

A fairly extensive set of count statements are checked, including
triggers, hits, crossings, enclosed, reaction counts, and molecule
counts.