Update test to work around Windows problem.
This test used to use COMPARE_COUNTS, but, since Xs are now used to
indicate an object is invalid/undefined, it became necessary to use
DIFF_FILE_CONTENT. However, it turns out this is a problem in Windows,
because the times are formatted differently (e.g. 1e-06 vs 1e-006). To
work around this problem, we can just output the iterations instead of
times by using ITERATION_LIST.