Modified the DG rxn example.
The previous version did an exact comparison of the DG and non-DG mdls.
They happened to exactly match, because of how the RNG was used. This is
no longer the case, so now we do a statistical comparison to make sure
that the DG case approximately matches the non-DG case in a manner
similar to what is done with expected_value_uni (i.e. it falls within a
reasonable range (3*std) of the average of the non-DGs average (using
100 seeds in this case)).