Add "sliver_leak" test for DGs.
This test is very similar to the normal sliver_leak test, but it uses dynamic
geometries as well.

Note: The molecules appear to be contained within the mesh as indicated
by the viz data, but the counts are wrong. As such, the test currently
fails and the cause is being investigated.