Dynamic geometry: add growing-shrinking box test
This is based off of an example that Bob created involving a growing and
shrinking box that contains diffusing volume and surface molecules.
Previously, this test would leak molecules, but it seems to be working
as of commmit a663a1b5 in the dynamic_meshes branch of the MCell repo.