Fix import paths to reference local copies.
I changed all the import paths to point to the mcellteam repo
instead of the haskelladdict one. NOTE: If this would have been
a real fork then changing the imports have been a hack and it would
have been much better to make this repo a remote of the original.
However, since the mcellteam repo is now the main repo, changing
the imports is probably fine.