Verify expected value of bimolecular reactions
This test checks the molecule counts from a number of different bimolecular
reactions throughout the simulation to make sure it falls within a reasonable
range of the expected value. The expected value is based off the mean molecule
counts from 1000 MCell simulations (which is comparable to results from an
ODE).

This test is similar to the expected_value_uni test.

Surface-surface tests still need to be added.