Switch test file format from json to toml
The following commit removes json as spec for our test files and
replaces it by toml. The reason for toml is readability and
simplicity. The available toml parser for go is more or less
a drop-in replacement for the previous json parser tipping the scale
in favor of toml versus yaml.

NOTE 1: In order to use nutmeg's new toml support two manual steps
are required:

1) Install the toml parser: go get github.com/BurntSushi/toml
2) Convert your nutmeg.conf from json to toml by simply removing
the top and bottom curly braces.

After that, nutmeg should work as before.

NOTE 2: The json test files were auto-converted to toml (with a few
small hand edits) and are thus not as pretty as they could be.
E.g., the "hacked" json comments via a description key-value pair
should be replaced by actual comments, etc.. Hopefully, we can
continue improving the new test_description.toml files as we go.