Test that SMs don't end up outside of PBs.
This test makes sure that surface molecules don't end up on portions of
surfaces that are outside of the periodic boundaries. Molecules are
released on a plane that cuts through the periodic boundaries. Part of
the portion outside of the PBs has a surface region called nothinghere.
If molecules ever end up on nothinghere, something has gone seriously
wrong.