Add test for recent MCell bugfix.
This test addresses a bug that was fixed in MCell commit d606a65f. This
bug was caused by changes to the surface molecule placement code when
using non-traditional PBCs. Specifically, the test verifies that a
surface molecule product can successfully engage in a bimolecular
reaction with a volume molecule. It should be noted that this bug also
affected non-PBC models.